Changed naming of tests
