+++ src/ServiceControl.AcceptanceTests/MultiInstance/When_processed_message_multi_instance_searched_by_messagetype.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Reflection;
     using System.Security.Cryptography;
     using System.Text;
@@ -16,8 +17,9 @@
     using ServiceBus.Management.AcceptanceTests.Contexts;
     using ServiceBus.Management.Infrastructure.Settings;
     using ServiceControl.CompositeViews.Messages;
+    using ServiceControl.Infrastructure.Settings;
 
-    public class When_messages_have_been_successfully_processed_by_multiple_instances_3 : AcceptanceTest
+    public class When_processed_message_multi_instance_searched_by_messagetype : AcceptanceTest
     {
         private const string Master = "master";
         private static string AuditMaster = $"{Master}.audit";
@@ -30,18 +32,38 @@
         private string addressOfRemote;
 
         [Test]
-        public void Should_be_found_in_search_by_messageId()
+        public void Should_be_found()
         {
             SetInstanceSettings = ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues;
 
             var context = new MyContext();
-            List<MessagesView> response;
+            List<MessagesView> response = new List<MessagesView>();
+
+            //search for the message type
+            var searchString = typeof(MyMessage).Name;
 
             Define(context, Remote1, Master)
-                .WithEndpoint<Sender>(b => b.Given((bus, c) => { bus.Send(new MyMessage()); }))
+                .WithEndpoint<Sender>(b => b.Given((bus, c) =>
+                {
+                    bus.Send(new MyMessage());
+                    bus.SendLocal(new MyMessage());
+                }))
                 .WithEndpoint<ReceiverRemote>()
-                .Done(c => c.Remote1MessageId != null && TryGetMany("/api/messages/search/" + c.Remote1MessageId, out response, instanceName: Master))
+                .Done(c => TryGetMany("/api/messages/search/" + searchString, out response, instanceName: Master) && response.Count == 2)
                 .Run(TimeSpan.FromSeconds(40));
+
+            var expectedMasterInstanceId = InstanceIdGenerator.FromApiUrl(SettingsPerInstance[Master].ApiUrl);
+            var expectedRemote1InstanceId = InstanceIdGenerator.FromApiUrl(SettingsPerInstance[Remote1].ApiUrl);
+
+            var masterMessage = response.SingleOrDefault(msg => msg.MessageId == context.MasterMessageId);
+
+            Assert.NotNull(masterMessage, "Master message not found");
+            Assert.AreEqual(expectedMasterInstanceId, masterMessage.InstanceId, "Master instance id mismatch");
+
+            var remote1Message = response.SingleOrDefault(msg => msg.MessageId == context.Remote1MessageId);
+
+            Assert.NotNull(remote1Message, "Remote1 message not found");
+            Assert.AreEqual(expectedRemote1InstanceId, remote1Message.InstanceId, "Remote1 instance id mismatch");
         }
 
         private void ConfigureRemoteInstanceForMasterAsWellAsAuditAndErrorQueues(string instanceName, Settings settings)
@@ -76,6 +98,23 @@
                     .AuditTo(Address.Parse(AuditMaster))
                     .ErrorTo(Address.Parse(ErrorMaster))
                     .AddMapping<MyMessage>(typeof(ReceiverRemote));
+            }
+
+            public class MyMessageHandler : IHandleMessages<MyMessage>
+            {
+                public MyContext Context { get; set; }
+
+                public IBus Bus { get; set; }
+
+                public ReadOnlySettings Settings { get; set; }
+
+                public void Handle(MyMessage message)
+                {
+                    Context.EndpointNameOfReceivingEndpoint = Settings.EndpointName();
+                    Context.MasterMessageId = Bus.CurrentMessageContext.Id;
+
+                    Thread.Sleep(200);
+                }
             }
         }
 
@@ -147,6 +186,7 @@
 
         public class MyContext : ScenarioContext
         {
+            public string MasterMessageId { get; set; }
             public string Remote1MessageId { get; set; }
 
             public string EndpointNameOfReceivingEndpoint { get; set; }
