+++ src/ServiceControl.AcceptanceTests/ServiceControl.AcceptanceTests.csproj
@@ -271,15 +271,15 @@
     <Compile Include="MultiInstance\When_a_message_retry_audit_is_sent_to_a_remote_instance.cs" />
     <Compile Include="MultiInstance\When_endpoint_detected_via_audits_on_slave.cs" />
     <Compile Include="MultiInstance\When_issuing_retry_on_master.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_1.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_2.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_3.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_4.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_5.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_6.cs" />
+    <Compile Include="MultiInstance\When_processed_message_multi_instance_by_messages.cs" />
+    <Compile Include="MultiInstance\When_processed_message_multi_instance_searched_by_messagetype.cs" />
+    <Compile Include="MultiInstance\When_processed_message_multi_instance_searched_by_messageid.cs" />
+    <Compile Include="MultiInstance\When_processed_message_multi_instance_searched_by_conversationId.cs" />
+    <Compile Include="MultiInstance\When_proessed_message_multi_instance_endpoint_by_messages.cs" />
+    <Compile Include="MultiInstance\When_processed_message_multi_instance_endpoint_by_messagetype.cs" />
     <Compile Include="MultiInstance\When_multiple_messages_are_emitted_by_a_saga.cs" />
     <Compile Include="MultiInstance\When_remote_instance_is_not_reachable.cs" />
-    <Compile Include="MultiInstance\When_messages_have_been_successfully_processed_by_multiple_instances_7.cs" />
+    <Compile Include="MultiInstance\When_processed_message_multi_instance_endpoint_processing.cs" />
     <Compile Include="MultiInstance\When_requesting_a_message_body_on_master.cs" />
     <Compile Include="Recoverability\Groups\When_a_group_is_retried.cs" />
     <Compile Include="Recoverability\Groups\When_a_group_is_archived.cs" />
